+++ e2e/e2e_test.go
@@ -32,6 +32,9 @@ import (
 	"github.com/k8snetworkplumbingwg/whereabouts/pkg/iphelpers"
 	wbstorage "github.com/k8snetworkplumbingwg/whereabouts/pkg/storage/kubernetes"
 	"github.com/k8snetworkplumbingwg/whereabouts/pkg/types"
+
+	// Import node slice tests to execute in the suite
+	_ "github.com/k8snetworkplumbingwg/whereabouts/e2e/e2e_node_slice"
 )
 
 const (
